% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_patterns.R
\name{analyze_issue_correlations}
\alias{analyze_issue_correlations}
\title{Correlate Issues with Data Characteristics}
\usage{
analyze_issue_correlations(test_results, cohort_data)
}
\arguments{
\item{test_results}{Result from run_dichotomous_tests()}

\item{cohort_data}{Original cohort data frame}
}
\value{
List with correlations
}
\description{
Correlate Issues with Data Characteristics
}
