% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{batch_log_issues}
\alias{batch_log_issues}
\title{Batch Log Issues from Test Results}
\usage{
batch_log_issues(test_results)
}
\arguments{
\item{test_results}{List of test results from run_dichotomous_tests()}
}
\value{
Summary of logged issues
}
\description{
Process multiple test results and log all issues.
}
