% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_new_methods.R
\name{bootstrap_validation}
\alias{bootstrap_validation}
\title{Bootstrap Validation}
\usage{
bootstrap_validation(
  df,
  methods = list(REML = function(d) test_reml_basic(d), CR2 = function(d)
    test_rve_cr2(d), Adaptive = function(d) adaptive_rve(d)),
  n_bootstrap = 500,
  seed = 123
)
}
\arguments{
\item{df}{Original data frame}

\item{methods}{Named list of methods to validate}

\item{n_bootstrap}{Number of bootstrap samples (default: 500)}

\item{seed}{Random seed}
}
\value{
Bootstrap validation results
}
\description{
Validate methods using bootstrap resampling to estimate variability.
}
