% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_issues.R
\name{categorize_issue}
\alias{categorize_issue}
\title{Categorize Issue by Type}
\usage{
categorize_issue(issue_report)
}
\arguments{
\item{issue_report}{Report from detect_all_issues()}
}
\value{
Character vector of issue categories
}
\description{
Categorize Issue by Type
}
