% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_rve.R
\name{compare_rve_methods}
\alias{compare_rve_methods}
\title{Compare RVE Methods}
\usage{
compare_rve_methods(df, cluster_var = "review_id")
}
\arguments{
\item{df}{Data frame with meta-analysis data}

\item{cluster_var}{Variable to use for clustering}
}
\value{
Data frame with comparison results
}
\description{
Run all RVE variants and compare performance.
}
