% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohort_tests.R
\name{create_downsampled_cohort}
\alias{create_downsampled_cohort}
\title{Create Cohort Subsample for Quick Testing}
\usage{
create_downsampled_cohort(full_df, n_analyses = 50, seed = 123)
}
\arguments{
\item{full_df}{Full cohort data frame}

\item{n_analyses}{Number of analyses to sample}

\item{seed}{Random seed}
}
\value{
Subsampled data frame
}
\description{
Create Cohort Subsample for Quick Testing
}
