% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohort_tests.R
\name{generate_test_report}
\alias{generate_test_report}
\title{Generate Test Report}
\usage{
generate_test_report(test_results, output_path = NULL)
}
\arguments{
\item{test_results}{Result from run_dichotomous_tests()}

\item{output_path}{Path to save report (NULL to print)}
}
\value{
Summary statistics
}
\description{
Generate Test Report
}
