% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{get_critical_issues}
\alias{get_critical_issues}
\title{Get Critical Issues}
\usage{
get_critical_issues()
}
\value{
Data frame of critical and high severity issues
}
\description{
Get Critical Issues
}
