% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{get_issues_by_method}
\alias{get_issues_by_method}
\title{Get Issues by Method}
\usage{
get_issues_by_method(method)
}
\arguments{
\item{method}{Method name (e.g., "REML", "RVE-CR2")}
}
\value{
Data frame of issues for the specified method
}
\description{
Get Issues by Method
}
