% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{get_issues_by_type}
\alias{get_issues_by_type}
\title{Get Issues by Type}
\usage{
get_issues_by_type(issue_type)
}
\arguments{
\item{issue_type}{Issue type (e.g., "convergence", "variance")}
}
\value{
Data frame of issues of the specified type
}
\description{
Get Issues by Type
}
