% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_registry.R
\name{get_method_recommendations}
\alias{get_method_recommendations}
\title{Get method recommendations based on data characteristics}
\usage{
get_method_recommendations(
  n_effects,
  n_clusters,
  outcome_type = "DICH",
  has_dependencies = TRUE
)
}
\arguments{
\item{n_effects}{Number of effect sizes in the analysis}

\item{n_clusters}{Number of independent clusters}

\item{outcome_type}{Type of outcome ("DICH", "CONT", "GENSUM")}

\item{has_dependencies}{Whether there are dependent effect sizes}
}
\value{
Character vector of recommended methods
}
\description{
Get method recommendations based on data characteristics
}
