% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_tau2.R
\name{get_tau2_recommendation}
\alias{get_tau2_recommendation}
\title{Get Recommended Tau² Method}
\usage{
get_tau2_recommendation(n_studies)
}
\arguments{
\item{n_studies}{Number of studies}
}
\value{
Recommended method and reasoning
}
\description{
Returns recommendation for tau² estimation based on sample size.
}
