% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_tau2.R
\name{improved_tau2}
\alias{improved_tau2}
\title{Improved Heterogeneity (tau²) Estimation}
\description{
Better methods for estimating between-study heterogeneity in multilevel
meta-analysis, especially for small samples and sparse data.
}
