% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_tau2.R
\name{improved_tau2_estimator}
\alias{improved_tau2_estimator}
\title{Improved Tau² Estimator}
\usage{
improved_tau2_estimator(yi, vi, method = "REML")
}
\arguments{
\item{yi}{Vector of effect sizes}

\item{vi}{Vector of sampling variances}

\item{method}{Primary estimation method ("REML", "DL", "SJ", "PM")}
}
\value{
List with tau² estimate and method details
}
\description{
Computes tau² using a combination of estimators with small-sample
corrections.
}
