% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{issue_tracker}
\alias{issue_tracker}
\title{Issue Tracking System for Meta-Analysis Testing}
\description{
Functions to log, track, and report issues found during meta-analysis
method testing on Cochrane review datasets.
}
