% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{load_issue_database}
\alias{load_issue_database}
\title{Load Issue Database from File}
\usage{
load_issue_database(file_path = "issues_database.rda")
}
\arguments{
\item{file_path}{Path to the .rda file}
}
\value{
Invisible TRUE
}
\description{
Load Issue Database from File
}
