% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_tau2.R
\name{qe_tau2_estimator}
\alias{qe_tau2_estimator}
\title{QE-based Tau² Estimator}
\usage{
qe_tau2_estimator(yi, vi)
}
\arguments{
\item{yi}{Vector of effect sizes}

\item{vi}{Vector of sampling variances}
}
\value{
List with tau² estimate
}
\description{
Uses the Q-statistic to estimate tau², which can be more stable
in small samples.
}
