% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohort_tests.R
\name{run_cohort_tests}
\alias{run_cohort_tests}
\title{Run Cohort Tests on Dichotomous Outcomes}
\description{
Functions to run systematic tests on the dichotomous outcomes (logOR) cohort
from the MLM501 Cochrane reviews dataset.
}
