% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohort_tests.R
\name{run_dichotomous_tests}
\alias{run_dichotomous_tests}
\title{Run Dichotomous Tests on All Analyses}
\usage{
run_dichotomous_tests(
  mlm_data_path = NULL,
  max_analyses = NULL,
  subsample_fraction = NULL,
  methods = c("REML", "RVE-CR2", "RVE-CR1"),
  progress = TRUE
)
}
\arguments{
\item{mlm_data_path}{Path to mlm_effects.csv}

\item{max_analyses}{Maximum number of analyses to test (NULL for all)}

\item{subsample_fraction}{Fraction of analyses to randomly sample (NULL for all)}

\item{methods}{Methods to test}

\item{progress}{Show progress bar}
}
\value{
List of all test results
}
\description{
Run Dichotomous Tests on All Analyses
}
