% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohort_tests.R
\name{run_single_analysis_tests}
\alias{run_single_analysis_tests}
\title{Run Tests on a Single Analysis}
\usage{
run_single_analysis_tests(df, methods = c("REML", "RVE-CR2", "RVE-CR1"))
}
\arguments{
\item{df}{Data frame for a single analysis_id}

\item{methods}{Methods to test}
}
\value{
Result list
}
\description{
Run Tests on a Single Analysis
}
