% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tracker.R
\name{save_issue_database}
\alias{save_issue_database}
\title{Save Issue Database to File}
\usage{
save_issue_database(file_path = "issues_database.rda")
}
\arguments{
\item{file_path}{Path to save the .rda file}
}
\value{
Invisible TRUE
}
\description{
Save Issue Database to File
}
