% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_new_methods.R
\name{simulation_study}
\alias{simulation_study}
\title{Simulation Study for Method Validation}
\usage{
simulation_study(
  scenarios,
  methods = list(REML = function(d) test_reml_basic(d), CR2 = function(d)
    test_rve_cr2(d)),
  n_sim = 1000,
  seed = 123
)
}
\arguments{
\item{scenarios}{List of simulation scenarios}

\item{methods}{Methods to test}

\item{n_sim}{Number of simulations per scenario (default: 1000)}

\item{seed}{Random seed}
}
\value{
Simulation results
}
\description{
Run simulation studies to validate method properties under known conditions.
}
