% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_new_methods.R
\name{summarize_validation}
\alias{summarize_validation}
\title{Summarize Validation Results}
\usage{
summarize_validation(results, methods, metrics)
}
\arguments{
\item{results}{Results from validate_method_comparison()}

\item{methods}{Method names}

\item{metrics}{Metrics to compute}
}
\value{
Summary data frame
}
\description{
Summarize Validation Results
}
