% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{test_all_methods}
\alias{test_all_methods}
\title{Test All Methods on a Single Dataset}
\usage{
test_all_methods(
  df,
  cluster_var = "review_id",
  methods = c("REML", "RVE-CR2", "RVE-CR1"),
  silent = FALSE
)
}
\arguments{
\item{df}{Data frame with meta-analysis data}

\item{cluster_var}{Variable to use for clustering}

\item{methods}{Character vector of methods to test ("all" for all available)}

\item{silent}{Suppress messages}
}
\value{
Data frame with results for all methods
}
\description{
Test All Methods on a Single Dataset
}
