% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{test_rve_cr2}
\alias{test_rve_cr2}
\title{Test RVE-CR2 Method (clubSandwich)}
\usage{
test_rve_cr2(df, cluster_var = "review_id", silent = FALSE)
}
\arguments{
\item{df}{Data frame with meta-analysis data}

\item{cluster_var}{Variable to use for clustering}

\item{silent}{Suppress messages}
}
\value{
List with fit object, convergence status, warnings, timing
}
\description{
Test RVE-CR2 Method (clubSandwich)
}
